The hashing functions are more generally useful; move to better spot
